# PERF: nanops