@alimcmaster1 2021-08-31
LGTM. Question how come we only apply this to certain nanops. Why not `nanstd` for example?