@jbrockmendel 2021-08-31
> Question how come we only apply this to certain nanops

These are the only ones that showed a big difference in the asvs.