@jbrockmendel 2021-11-21
> add some threshold for the shape

seems reasonable