@jorisvandenbossche 2021-11-21
Added that in https://github.com/pandas-dev/pandas/pull/44566